Working local core apps (traefik and portainer) but not app yet.
Found a beautiful example with proper explanations, going with
it. Link to blog is in README file and in project's journal.

Commented out https stuff, as doing it locally is more difficult than
just doing it online with automatic cert retrieval.

Also moved sensitive traefik data to an untracked file and made sure
it works (that passwords defined in it can be used to log in).

Also added ignoring of portainer data and the sensitive configuration
to .gitignore.