Condolidate two directories used for local development into one
Because of issues with Syncthing (that were later resolved using
dot-ignore file) I needed separate dirs to separate Vagrant
directories on different machines.

This is not outrightly obvious for an outsider looking at the project
and was therefore consolidated into one.

To ensure no work is lost and that the dragon-machines dir contained
everything, a 'dir -ur dragon-machines talos-machines' command was run
to see the differences between these folders.